+++ .env.example
@@ -1,2 +1,6 @@
 DB_HOST=
 SECRET_KEY=
+
+SENDGRID_API_KEY=
+SITE_NAME=
+SENDER_MAIL=
