+++ controllers/contacts/index.js
@@ -1,0 +1,15 @@
+const getAll = require('./getAll');
+const getById = require('./getById');
+const addContact = require('./addContact');
+const deleteById = require('./deleteById');
+const updateById = require('./updateById');
+const updateFavorite = require('./updateFavorite');
+
+module.exports = {
+  getAll,
+  getById,
+  addContact,
+  deleteById,
+  updateById,
+  updateFavorite,
+};
