+++ controllers/users/getUserByToken.js
@@ -1,0 +1,11 @@
+const getUserByToken = async (req, res) => {
+  const { email, subscription } = req.user;
+  res.json({
+    user: {
+      email,
+      subscription,
+    },
+  });
+};
+
+module.exports = getUserByToken;
