+++ helpers/index.js
@@ -1,0 +1,3 @@
+const sendMail = require('./sendMail');
+
+module.exports = { sendMail };
