+++ models/user.js
@@ -25,6 +25,14 @@ const userSchema = Schema(
       type: String,
       default: null,
     },
+    verify: {
+      type: Boolean,
+      default: false,
+    },
+    verificationToken: {
+      type: String,
+      required: [true, 'Verify token is required'],
+    },
     avatarURL: { type: String, default: '' },
   },
   { versionKey: false, timestamps: true },
