+++ package.json
@@ -10,6 +10,7 @@
     "lint:fix": "eslint --fix **/*.js"
   },
   "dependencies": {
+    "@sendgrid/mail": "^7.6.0",
     "bcrypt": "^5.0.1",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
@@ -22,7 +23,8 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.3",
     "morgan": "1.10.0",
-    "multer": "^1.4.4"
+    "multer": "^1.4.4",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "eslint": "7.19.0",
